add casts to long when multiplying so one of the values isn't truncated

+++ src/main/java/travis_ci_tutorial_java/SimpleCalculator.java
@@ -10,9 +10,9 @@ public class SimpleCalculator {
 	}
 	
 	public int multiply(int a, int b) {
-		long check = a * b;
+		long check = (long) a * (long) b;
 		int ret = a * b;
-		if(check != ret) {
+		if(check != (long) ret) {
 			if(check > 0) {
 				throw new ArithmeticException("Integer Overflow Detected");
 			} else {
